refactor: move firm construction to separate crate
