+++ compiler-lib/Cargo.toml
@@ -21,6 +21,7 @@ lexer = { path = "../lexer" }
 parser = { path = "../parser" }
 utils = { path = "../utils" }
 type-checking = { path = "../type_checking" }
+firm-construction = { path = "../firm-construction" }
 failure = "0.1.3"
 termcolor = "1.0.4"
 derive_more = "0.13"
