+++ firm-construction/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "firm-construction"
+version = "0.1.0"
+authors = [
+    "flip1995 <>",
+    "hediet <>",
+    "joshuabach <>",
+    "problame <>",
+    "reiner-dolp <>"
+]
+edition = "2018"
+license = "MIT OR Apache-2.0"
+
+[dependencies]
+failure = "0.1.3"
+compiler-shared = { path = "../compiler-shared" }
+utils = { path = "../utils" }
+asciifile = { path = "../asciifile" }
+strtab = { path = "../strtab" }
+type-checking = { path = "../type_checking" }
+parser = { path = "../parser" }
+
+#diagnostics = { path = "../diagnostics" }
+libfirm-rs = { path = "../libfirm-rs" }
+libc = "0.2"
+#lexer = { path = "../lexer" }
+derive_more = "0.13"
+strum = "0.11.0"
+strum_macros = "0.11.0"
+itertools = "0.7"
+log = "0.4"
+lazy_static = "1.2.0"
